Add the Calculate_ICS function in  arviz/utils.py
